:recycle: Refactor assignment detail view into component.
