+++ src/components/AssignmentDetail.vue
@@ -1,43 +1,80 @@
 <template>
-  <div class="flex-col detail">
-    <div class="name title">
-      <p>{{ assignment.name }}</p>
+    <div class="grid">
+      <div class="full-left-grid border">
+        <p><md-icon class="md-size-2x">assignment</md-icon></p>
+        <p class="md-display-3">{{ assignment.name }}</p>
+        <div
+          v-if="assignment.tag"
+          @click="navigateToTagPage(assignment.tag.id)"
+        >
+          <TagBoxChip
+            :color="assignment.tag.color"
+            :text="assignment.tag.name"
+            class="tag"
+          ></TagBoxChip>
+        </div>
+        <p class="md-subheading">Due Date: {{ this.getFormatDate }}</p>
+        <p class="md-subheading">Submission Time: {{ this.getFormatTime }}</p>
+      </div>
+      <div class="upper-right-grid">
+        <detail-card :title="detailCardTexts.author">
+          <slot name="authorPicture"></slot>
+        </detail-card>
+        <detail-card :title="detailCardTexts.description">
+          <div v-if="assignment.description">
+            <p class="content">{{ assignment.description }}</p>
+          </div>
+          <div v-else>
+            <md-icon>minimize</md-icon>
+          </div>
+
+        </detail-card>
+        <detail-card :title="detailCardTexts.reminderTime">
+          <div
+            v-if="
+              assignment.reminderTime && assignment.reminderTime.length !== 0
+            "
+          >
+            <md-chip
+              v-for="time in assignment.reminderTime" 
+              :key="time"
+              :style="`background-color: ${assignment.tag.color} !important; color: ${getFontColor(
+                assignment.tag.color
+              )} !important`"
+              ><md-icon>notifications</md-icon>{{ getHumanReadableTime(time) }}</md-chip
+            >
+          </div>
+          <div v-else>
+            <md-icon>minimize</md-icon>
+          </div>
+        </detail-card>
+      </div>
+      <div class="lower-right-grid">
+        <slot name="buttons"></slot>
+      </div>
     </div>
-    <div
-      v-if="assignment.tag"
-      class="tag"
-      @click="navigateToTagPage(assignment.tag.id)"
-    >
-      <tag-box :tag="assignment.tag"></tag-box>
-    </div>
-    <div class="description text-gray">
-      <p class="header">Description:</p>
-      <p class="content">{{ assignment.description }}</p>
-    </div>
-    <div v-if="this.assignment.dueDate" class="due-date text-gray">
-      <p>Due Date:</p>
-      <p class="formatted">{{ getFormatDate }}</p>
-      <p>Submission Time:</p>
-      <p class="formatted">{{ getFormatTime }}</p>
-    </div>
-    <div v-else class="due-date text-gray">
-      <p>No Due Date</p>
-    </div>
-  </div>
 </template>
 
 <script lang="ts">
 import { Component, Prop, Vue } from "vue-property-decorator";
 import { Assignment } from "@/types/garnbarn/Assignment";
-import TagBox from "@/components/Tag/TagBox.vue";
+import TagBoxChip from "@/components/Tag/TagBoxChip.vue";
+import DetailCard from "@/components/DetailCard.vue";
 
 @Component({
   components: {
-    TagBox,
+    TagBoxChip,
+    DetailCard,
   },
 })
 export default class AssignmentDetail extends Vue {
   @Prop({ required: true }) readonly assignment!: Assignment;
+
+  detailCardTexts = {
+    author: "Author: ",
+    description: "Description: ",
+    reminderTime: "Reminder Time: ",
+  };
 
   convertUnixTimeToDate(unix_timestamp: number): Date {
     var date = new Date(unix_timestamp);
@@ -67,6 +104,46 @@ export default class AssignmentDetail extends Vue {
     return "Unknown";
   }
 
+  getHumanReadableTime(unixTime: number): string {
+    var message = "";
+
+    var day = Math.floor(unixTime / 86400);
+    unixTime -= day * 86400;
+    if (day >= 1) {
+      message += day > 1 ? `${day} Days` : `${day} Day`;
+    }
+
+    var hour = Math.floor(unixTime / 3600) % 24;
+    unixTime -= hour * 3600;
+    if (hour >= 1) {
+      message += hour > 1 ? `${hour} hours` : `${hour} hour`;
+    }
+
+    var minute = Math.floor(unixTime / 60) % 60;
+    unixTime -= minute * 60;
+    if (minute >= 1) {
+      message += minute > 1 ? `${minute} minutes` : `${minute} minute`;
+    }
+
+    return message;
+  }
+
+  hexToRgb(hex: string): Array<number> {
+    const r = parseInt(hex.slice(1, 3), 16);
+    const g = parseInt(hex.slice(3, 5), 16);
+    const b = parseInt(hex.slice(5, 7), 16);
+    return [r, g, b];
+  }
+
+  getFontColor(hex: string): string {
+    const rgb = this.hexToRgb(hex);
+    if (rgb[0] * 0.299 + rgb[1] * 0.587 + rgb[2] * 0.114 > 186) {
+      return "#000000";
+    } else {
+      return "#ffffff";
+    }
+  }
+
   navigateToTagPage(tagId: number): void {
     this.$router.push(`/tag/${tagId}`);
   }
@@ -74,47 +151,34 @@ export default class AssignmentDetail extends Vue {
 </script>
 
 <style scoped>
-.detail {
-  padding: 2rem 5rem;
+.border {
+  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.5);
+  border-radius: 10px;
+  border: 1px solid black;
 }
 
-.flex-col {
-  flex: 1 1 0%;
+.grid {
+  margin: 2rem 0rem;
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 2rem;
+}
+
+.full-left-grid {
+  padding: 5rem;
+  grid-row: span 2 / span 2;
+  display: flex;
   flex-direction: column;
-  text-align: left;
+  justify-content: center;
+  align-items: center;
 }
 
-.title {
-  font-size: 2rem !important;
-  height: fit-content;
-}
-
-.description {
-  padding-right: 40%;
-}
-
-.description {
-  font-size: 1.5rem;
-}
-
-.description .content {
+.content {
   font-size: 1.2rem;
   overflow-wrap: break-word;
 }
 
-.due-date {
-  font-size: 1.3rem;
-}
-
-.due-date .formatted {
-  font-size: 1.2rem;
-}
-
-.text-gray {
-  color: #616161;
-}
-
-.tag {
-  cursor: pointer;
+.left-align {
+  text-align: left;
 }
 </style>
